feat: default product image

+++ sqlite/migrations/007-remove-field-quantity-product.ts
@@ -1,3 +1,5 @@
 export async function up(db: import("sqlite").Database) {
+  if (process.env.NODE_ENV !== "production") return Promise.resolve();
+
   await db.exec(`ALTER TABLE products DROP COLUMN quantity;`);
 }
